css & connexion.jsx modified
