+++ frontend/src/App.jsx
@@ -6,7 +6,6 @@ import Home from "./pages/Home";
 import Connexion from "./pages/Connexion";
 import Subscribe from "./pages/Subscribe";
 import Offers from "./pages/Offers";
-import AdminConnexion from "./pages/AdminConnexion";
 import Dashboard from "./pages/Dashboard";
 
 // Import des composants
@@ -34,7 +33,6 @@ function App() {
                 <Route path="connexion" element={<Connexion />} />
                 <Route path="subscribe" element={<Subscribe />} />
                 <Route path="offers" element={<Offers />} />
-                <Route path="admin-connexion" element={<AdminConnexion />} />
                 <Route path="dashboard/*" element={<Dashboard />} />
               </Routes>
             </Router>
