--- frontend/src/css/components/ConnexionLinks.css
@@ -1,41 +1,0 @@
-.nav-bar-connexion {
-  display: flex;
-  justify-content: center;
-  margin: 10vh auto 10vh auto;
-}
-
-.nav-bar-connexion li {
-  display: inline-block;
-}
-
-.login-selected a {
-  font-weight: bolder;
-  font-size: larger;
-  text-decoration: underline 2px;
-  text-underline-offset: 8px;
-  color: black;
-}
-
-.signin-selected a {
-  color: var(--basic-grey);
-  text-decoration: none;
-  margin: auto 40px auto 20px;
-  font-size: larger;
-}
-
-#titleEspaceAdmin {
-  font-weight: bolder;
-  font-size: 28px;
-  text-decoration: underline 2px;
-  text-underline-offset: 8px;
-  color: black;
-  margin: 4rem auto;
-  text-align: center;
-}
-
-@media screen and (min-width: 768px) {
-  .login-selected a,
-  .signin-selected a {
-    font-size: 28px;
-  }
-}
