--- frontend/src/css/components/FormularConnexion.css
@@ -1,12 +1,0 @@
-.error-email,
-.error-password {
-  color: var(--main-color);
-  font-style: italic;
-  font-size: smaller;
-  position: absolute;
-  width: max-content;
-  margin-top: 4rem;
-}
-
-@media screen and (min-width: 768px) {
-}
