--- frontend/src/css/pages/AdminConnexion.css
@@ -1,16 +1,0 @@
-.container-admin-connexion {
-  background-image: url("../../assets/images/admin-connexion.png");
-  background-size: 50%;
-  background-repeat: no-repeat;
-  background-position: right;
-  background-color: var(--white-80);
-  border: solid 1px var(--main-color);
-  box-shadow: var(--shadow_initial);
-  border-radius: 16px;
-  width: 60vw;
-  height: 60vh;
-  margin: 5rem auto;
-  display: flex;
-  padding-left: 4rem;
-  padding-top: 4rem;
-}
